+++ python/py_src/sudachipy/__init__.py
@@ -1,5 +1,6 @@
 from .sudachipy import (
     Dictionary,
+    TextNormalizer,
     Tokenizer,
     SplitMode,
     MorphemeList,